add function calls
The function definitions must start from the beginning of the string,
and '!' is used to start a function definition.
The '!!' marks the end of the function definition section and beginning of the 'main' code.
After that one should not have any !-marks.

To call the function, use "C", which operates on the value.

An example program string that signals SOS with red and green:

(ILED:!X!XR10KP X30KP!XG30KP X10KP!X100KP!1C1C1C!2C2C2C!!4C5C4C3C)